+++ examples/adapters/pg_client.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'bundler/setup'
-require 'polyphony/postgres'
+require 'polyphony/adapters/postgres'
 
 def get_records
   $db.query('select 1 as test')