+++ lib/polyphony/adapters/postgres.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../polyphony'
+require_relative '../../polyphony'
 require 'pg'
 
 # PG overrides